move and rename _setup_picks to _check_info_inv
